+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 2021.6.7 (22 June)
+
+### Fixed
+
+* The CatManga source now parses manga descriptions properly.
+  Implicit wrapping of `str` descriptions in `models.Metadata` to dictionaries caused descriptions to be nested.
+
 ## 2021.6.6 (20 June)
 
 ### Added